Fix frame discarding behavior on timestamp mismatch.
The old method was flakey for the MicroMax system because
each trigger indicates the end of the frame, and we were
pretending that each trigger was the start of the frame by
caching the last two triggers. The behavior of discarding
triggers then depended implicitly on the timing of the mismatch
relative to the download and trigger.

This fix removes this bogus assumption, acknowledging that the
triggers occur at different times in the two systems.
Further patches will need to move this code into the camera impl
and restore discarding the first frame for *all* cameras as
the first frame from the ProEM is bogus for unrelated reasons.